net: ipa: don't use "full pipeline clear"
The REGISTER_WRITE IPA immediate command has an option to ensure
certain stages of the hardware pipeline are cleared as part of
performing the command.  Commands normally specify that header
processing should be cleared, but when suspending endpoints, as well
as when recovering from a modem crash, we issue a REGISTER_WRITE
command requesting a "full" pipeline clear.

Qualcomm has suggested that the "full" clear is more than what is
required, and that it could even contribute to occasional hangs we
have seen when it is used.

So stop doing (or even supporting) full pipeline clear for the
REGISTER_WRITE immediate command, and always use HPS clear instead.

Fixes: 731c46edad82 ("soc: qcom: ipa: immediate commands")
Signed-off-by: Alex Elder <elder@linaro.org>