net: ipa: *do* stop TX on suspend (temporarily)
When we suspend IPA hardware due to inactivity, we don't want to
disable network transmit.  We want transmit requests to cause the
hardware to be resumed.  With that in mind, commit 09ee011cff881
("net: ipa: don't stop TX on suspend") stopped disabling transmit
on the modem TX endpoint when we suspend.

Suspending an endpoint leads to a request to quiesce all underlying
channel activity.  And doing that involves waiting for the "last"
allocated transaction to complete.  This implicitly assumes that
there *is* a last one, and that no more transactions will be
allocated until after the quiesce operation completes.

Unfortunately, without disabling transmit on the modem TX endpoint,
we violate this assumption in the quiesce operation.

Fix this by restoring the removed call to netif_stop_queue(), but
wake the queue again after the modem TX has been suspended.

Fixes: 09ee011cff881 ("net: ipa: don't stop TX on suspend")
Signed-off-by: Alex Elder <elder@linaro.org>