net: ipa: add monitor endpoint support
Add support for the IPA monitor endpoint.  When enabled, the monitor
endpoint on an IPA instance supplies replicas of all packets that
are handled by the hardware.  To reduce bandwidth requirements,
large packets (larger than 192 bytes, for example) are truncated.
And to reduce processing load, multiple packets (with their headers)
may be aggregated into a single receive buffer.  Each aggregated
packet received on the monitor endpoint is preceded by a fixed-size
DPL (data protocol logging) header that describes the packet.
The DPL header includes:
  - the 16-bit length of the (truncated) packet data
  - a 32-bit metadata value associated with the packet
  - a source endpoint index, identifying the packet sender
  - a 48-bit time-of-day counter
  - an 8-bit sequence number
  - and an 8-bit "opcode", which will have value 0x10

Like any RX endpoint, the monitor endpoint has a fixed size transfer
ring to hold pages into which received packets are written.  When
the endpoint is enabled, received packets are written into the next
available buffer in the transfer ring.  When a buffer from the
monitor endpoint arrives it is passed to ipa_monitor_rx() to be
handled.  If a packet arrives on the monitor endpoint when no buffer
is available, it is dropped by hardware.

Packets received on the monitor endpoint will made available to
user space through a file interface.  When that file is opened,
the monitor endpoint will be enabled; the endpoint is disabled
again when the file is closed.

Buffers holding received packets are held until user space has read
their contents via the file.  We have no control over how fast user
space consumes packets from the file interface, so we limit the
number of buffers waiting to be read to the number of entries in the
monitor endpoint's transfer ring.  If they are not read fast enough,
newer packets will automatically be dropped by hardware.

To do this, we delay replenishing a receive buffer until after an
already-received buffer has been consumed by user space.

Signed-off-by: Alex Elder <elder@linaro.org>