net: ipa: introduce /dev/qcom_ipa0_monitor
Create a new misc device to provide a file interface that supplies
packet data received on the IPA monitor endpoint.

The misc device is named "qcom_ipaX_monitor", where X is the IPA
instance number (most likely just 0).  It is implemented as a read-
only stream-like file (lseek is not supported and there is no notion
of the "current position" of the file when open).  The device file
will be opened exclusively; opening the underlying monitor endpoint
fails if it's already open.

We store a copy of the file pointer to determine whether the device
file is open (used at driver shutdown time).  We also stash the
monitor pointer as private data in the file when it's opened.  In
case we shut down the driver while the file was open, we reset that
stashed value to NULL.  The file operations check for that, and
return an error rather than assuming the monitor structure can be
used.

Signed-off-by: Alex Elder <elder@linaro.org>