net: ipa: record netdev in channel
Currently, the only endpoints that have a network device associated
with them are the modem TX and RX endpoints.  The endpoint netdev
pointer is only assigned for those endpoints, never for any of the
others.

Record the same netdev pointer in the channel underlying these
endpoints as well.  Eventually we should be able to remove the
netdev pointer from the endpoint structure.

Signed-off-by: Alex Elder <elder@linaro.org>