net: ipa: kill ipa_gsi_channel_tx_queued()
Now that the channel holds a network device pointer, we can avoid
having gsi_channel_tx_queued() call ipa_gsi_channel_tx_queued(), and
have it call netdev_sent_queue() directly.

This move makes it obvious that the queued transaction count isn't
actually needed, so we can get rid of that field in the channel
structure.  And in doing that, we find the channel->trans_count
field is set but not otherwise used, so that can go away too.

Signed-off-by: Alex Elder <elder@linaro.org>