net: ipa: check for overflow in ipa-shared memory size
In ipa_mem_init(), the size of the IPA-resident memory is determined
and assigned to a 32-bit mem_size field in the IPA structure.  It's
possible that the assigned value is too big to be represented by the
field.

Stash the resource size in a new local variable, and verify the
end of the IPA memory range will fit in the field before making the
assignment.

Signed-off-by: Alex Elder <elder@linaro.org>