net: ipa; fix some kernel-doc issues
If a function documented with kernel-doc comments returns a value,
the documentation should indicate what is returned.  If W=2 is used
in the kernel build, a warning is issued wherever this occurs.

There are six spots in the IPA code that trigger that warning.  Add
a comment to each to make those warnings go away.

Signed-off-by: Alex Elder <elder@linaro.org>