net: ipa: restrict IPA-resident memory address to 32 bits
We record the base address of the "ipa-mem" memory range specified
in Device Tree, but we don't actually use it.

Instead, in ipa_mem_config() we request from hardware the address
and size of that memory range, and use the value returned--storing
it in the 32-bit mem_offset field in the IPA structure.

Rather than just getting rid if the mem_addr field, add a check in
ipa_mem_init() to ensure the base address provided by hardware
matches the expected value provided in the DTB.

The hardware assumes the address fits in 32 bits, so add a check
to verify the value provided by the DTB is representable before
assigning it.

Change the type of the mem_addr field to be u32, and rearrange where
it's defined in the IPA strucure to avoid a hole in the structure.

Signed-off-by: Alex Elder <elder@linaro.org>