net: ipa: kill ipa->mem_addr
After recording the IPA-resident memory address provided by the DTB,
we proceed to ignore it once we've verified it matches the address
provided by the hardware.

Instead, just use the mem_offset field first to record the address
provided by hardware, then continue to use it once we've verified
it matches what was reported by hardware.

This makes the mem_addr field unused, so get rid of it.

Signed-off-by:: Alex Elder <elder@linaro.org>