net: toshiba: tc956x: Warning fixes (W=1)
Pointers to the tc956xmac_cm3_tamap structure are used in arguments in
"tc956xmac.h", but it's never declared.

The "xmac" variable in tc956xmac_hwtstamp_set() is defined and set
but never used thereafter, so get rid of it.

The "tcr_config" variable in tc956xmac_hwtstamp_set() is unnecessarily
set to zero.

(I'm signing off on this, but we agree this isn't going anywhere public.)

Signed-off-by: Alex Elder <elder@riscstar.com>