net: toshiba: tc956x: sparse warnings
In tc_fill_entry(), two values that are defined to be 32-bit big endian
are assigned to plain old u32 variables.

The proper way to handle this is to explicitly convert the big endian
value to be in CPU byte order.  But the code presumably works, and
there is a little code involving shifts below this, which I assume
does the equivalent (though it doesn't look quite right to me).
So just force a cast to the target type.

Anyway, get rid of the sparse warnings by adding a cast.

Signed-off-by: Alex Elder <elder@risctar.com>